Rename FileUploadBox to FileUploadForm
