+++ Source/DesignSystem/index.ts
@@ -22,7 +22,7 @@ export { Tooltip } from './atoms/Tooltip';
 export { InputMessages, OutputMessages, Terminal, TerminalConnect, TerminalStreams } from './atoms/Terminal';
 
 // Molecules
-export { FileUploadBox, FileUploadFormProps } from './molecules/FileUploadBox';
+export { FileUploadForm, FileUploadFormProps } from './molecules/FileUploadForm';
 export { Graph } from './molecules/Metrics/Graph';
 export { NavigationBar } from './molecules/NavigationBar';
 export { SelectCard } from './molecules/SelectCard';
