+++ Source/DesignSystem/molecules/FileUploadForm/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { FileUploadBox, FileUploadFormProps } from './FileUploadBox';
+export { FileUploadForm, FileUploadFormProps } from './FileUploadForm';
